# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StayBook Room Booking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  renter: "renter",
  owner: "owner",
  both: "both",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export type RoomType = (typeof RoomType)[keyof typeof RoomType];

export const RoomType = {
  private: "private",
  shared: "shared",
  studio: "studio",
  meeting: "meeting",
} as const;

export interface User {
  id: string;
  name: string;
  phone: string;
  role: UserRole;
  avatar?: string | null;
  createdAt: string;
}

export interface CreateUserRequest {
  name: string;
  phone: string;
  role: UserRole;
}

export interface Room {
  id: string;
  ownerId: string;
  ownerName: string;
  title: string;
  description: string;
  city: string;
  address: string;
  pricePerHour: number;
  pricePerDay: number;
  roomType: RoomType;
  amenities: string[];
  maxGuests: number;
  images: string[];
  isAvailable: boolean;
  rating?: number | null;
  totalReviews: number;
  createdAt: string;
}

export interface CreateRoomRequest {
  ownerId: string;
  ownerName: string;
  title: string;
  description: string;
  city: string;
  address: string;
  pricePerHour: number;
  pricePerDay: number;
  roomType: RoomType;
  amenities: string[];
  maxGuests: number;
  images: string[];
}

export interface UpdateRoomRequest {
  title?: string;
  description?: string;
  city?: string;
  address?: string;
  pricePerHour?: number;
  pricePerDay?: number;
  roomType?: RoomType;
  amenities?: string[];
  maxGuests?: number;
  images?: string[];
  isAvailable?: boolean;
}

export type OrderBookingType =
  (typeof OrderBookingType)[keyof typeof OrderBookingType];

export const OrderBookingType = {
  hourly: "hourly",
  daily: "daily",
} as const;

export interface Order {
  id: string;
  roomId: string;
  roomTitle: string;
  roomAddress: string;
  renterId: string;
  renterName: string;
  renterPhone: string;
  ownerId: string;
  checkIn: string;
  checkOut: string;
  bookingType: OrderBookingType;
  totalAmount: number;
  guests: number;
  specialRequests?: string | null;
  status: OrderStatus;
  createdAt: string;
}

export type CreateOrderRequestBookingType =
  (typeof CreateOrderRequestBookingType)[keyof typeof CreateOrderRequestBookingType];

export const CreateOrderRequestBookingType = {
  hourly: "hourly",
  daily: "daily",
} as const;

export interface CreateOrderRequest {
  roomId: string;
  renterId: string;
  renterName: string;
  renterPhone: string;
  ownerId: string;
  checkIn: string;
  checkOut: string;
  bookingType: CreateOrderRequestBookingType;
  totalAmount: number;
  guests: number;
  specialRequests?: string;
}

export interface UpdateOrderStatusRequest {
  status: OrderStatus;
}

export type ListRoomsParams = {
  ownerId?: string;
  city?: string;
};

export type ListOrdersParams = {
  renterId?: string;
  ownerId?: string;
  status?: OrderStatus;
};