# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StayBook Room Booking API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User management
  - name: rooms
    description: Room listings
  - name: orders
    description: Booking orders
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create or login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: User created or found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all available rooms
      parameters:
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
        - name: city
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a room listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

  /rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room by ID
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room listing
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomRequest"
      responses:
        "200":
          description: Room updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room listing
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Room deleted

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for a user
      parameters:
        - name: renterId
          in: query
          required: false
          schema:
            type: string
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a booking order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (accept/reject/cancel/complete)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    UserRole:
      type: string
      enum: [renter, owner, both]

    OrderStatus:
      type: string
      enum: [pending, accepted, rejected, cancelled, completed]

    RoomType:
      type: string
      enum: [private, shared, studio, meeting]

    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        avatar:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - role
        - createdAt

    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required:
        - name
        - phone
        - role

    Room:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        ownerName:
          type: string
        title:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        pricePerHour:
          type: number
        pricePerDay:
          type: number
        roomType:
          $ref: "#/components/schemas/RoomType"
        amenities:
          type: array
          items:
            type: string
        maxGuests:
          type: integer
        images:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        rating:
          type: number
          nullable: true
        totalReviews:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - ownerId
        - ownerName
        - title
        - description
        - city
        - address
        - pricePerHour
        - pricePerDay
        - roomType
        - amenities
        - maxGuests
        - images
        - isAvailable
        - totalReviews
        - createdAt

    CreateRoomRequest:
      type: object
      properties:
        ownerId:
          type: string
        ownerName:
          type: string
        title:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        pricePerHour:
          type: number
        pricePerDay:
          type: number
        roomType:
          $ref: "#/components/schemas/RoomType"
        amenities:
          type: array
          items:
            type: string
        maxGuests:
          type: integer
        images:
          type: array
          items:
            type: string
      required:
        - ownerId
        - ownerName
        - title
        - description
        - city
        - address
        - pricePerHour
        - pricePerDay
        - roomType
        - amenities
        - maxGuests
        - images

    UpdateRoomRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        pricePerHour:
          type: number
        pricePerDay:
          type: number
        roomType:
          $ref: "#/components/schemas/RoomType"
        amenities:
          type: array
          items:
            type: string
        maxGuests:
          type: integer
        images:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean

    Order:
      type: object
      properties:
        id:
          type: string
        roomId:
          type: string
        roomTitle:
          type: string
        roomAddress:
          type: string
        renterId:
          type: string
        renterName:
          type: string
        renterPhone:
          type: string
        ownerId:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        bookingType:
          type: string
          enum: [hourly, daily]
        totalAmount:
          type: number
        guests:
          type: integer
        specialRequests:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/OrderStatus"
        createdAt:
          type: string
      required:
        - id
        - roomId
        - roomTitle
        - roomAddress
        - renterId
        - renterName
        - renterPhone
        - ownerId
        - checkIn
        - checkOut
        - bookingType
        - totalAmount
        - guests
        - status
        - createdAt

    CreateOrderRequest:
      type: object
      properties:
        roomId:
          type: string
        renterId:
          type: string
        renterName:
          type: string
        renterPhone:
          type: string
        ownerId:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        bookingType:
          type: string
          enum: [hourly, daily]
        totalAmount:
          type: number
        guests:
          type: integer
        specialRequests:
          type: string
      required:
        - roomId
        - renterId
        - renterName
        - renterPhone
        - ownerId
        - checkIn
        - checkOut
        - bookingType
        - totalAmount
        - guests

    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
      required:
        - status