# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StayBook Room Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestBookingType } from "./createOrderRequestBookingType";

export interface CreateOrderRequest {
  roomId: string;
  renterId: string;
  renterName: string;
  renterPhone: string;
  ownerId: string;
  checkIn: string;
  checkOut: string;
  bookingType: CreateOrderRequestBookingType;
  totalAmount: number;
  guests: number;
  specialRequests?: string;
}