# lib/api-zod/src/generated/types/createUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StayBook Room Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface CreateUserRequest {
  name: string;
  phone: string;
  role: UserRole;
}