# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StayBook Room Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderBookingType } from "./orderBookingType";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  roomId: string;
  roomTitle: string;
  roomAddress: string;
  renterId: string;
  renterName: string;
  renterPhone: string;
  ownerId: string;
  checkIn: string;
  checkOut: string;
  bookingType: OrderBookingType;
  totalAmount: number;
  guests: number;
  specialRequests?: string | null;
  status: OrderStatus;
  createdAt: string;
}