# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StayBook Room Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomType } from "./roomType";

export interface Room {
  id: string;
  ownerId: string;
  ownerName: string;
  title: string;
  description: string;
  city: string;
  address: string;
  pricePerHour: number;
  pricePerDay: number;
  roomType: RoomType;
  amenities: string[];
  maxGuests: number;
  images: string[];
  isAvailable: boolean;
  rating?: number | null;
  totalReviews: number;
  createdAt: string;
}